Fix infinite loop in assemble_excerpts
