+++ crates/edit_prediction_context/src/assemble_excerpts.rs
@@ -61,8 +61,8 @@ pub fn assemble_excerpts(
                                 buffer,
                                 &mut outline_ranges,
                             );
-                            child_outline_ix += 1;
                         }
+                        child_outline_ix += 1;
                     }
                 }
             }
@@ -157,168 +157,5 @@ pub fn merge_ranges(ranges: &mut Vec<Range<Point>>) {
         } else {
             index += 1;
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use gpui::{TestAppContext, prelude::*};
-    use indoc::indoc;
-    use language::{Buffer, Language, LanguageConfig, LanguageMatcher, OffsetRangeExt};
-    use pretty_assertions::assert_eq;
-    use std::{fmt::Write as _, sync::Arc};
-    use util::test::marked_text_ranges;
-
-    #[gpui::test]
-    fn test_rust(cx: &mut TestAppContext) {
-        let table = [
-            (
-                indoc! {r#"
-                    struct User {
-                        first_name: String,
-                        «last_name»: String,
-                        age: u32,
-                        email: String,
-                        create_at: Instant,
-                    }
-
-                    impl User {
-                        pub fn first_name(&self) -> String {
-                            self.first_name.clone()
-                        }
-
-                        pub fn full_name(&self) -> String {
-                    «        format!("{} {}", self.first_name, self.last_name)
-                    »    }
-                    }
-                "#},
-                indoc! {r#"
-                    struct User {
-                        first_name: String,
-                        last_name: String,
-                    …
-                    }
-
-                    impl User {
-                    …
-                        pub fn full_name(&self) -> String {
-                            format!("{} {}", self.first_name, self.last_name)
-                        }
-                    }
-                "#},
-            ),
-            (
-                indoc! {r#"
-                    struct «User» {
-                        first_name: String,
-                        last_name: String,
-                        age: u32,
-                    }
-
-                    impl User {
-                        // methods
-                    }
-                    "#
-                },
-                indoc! {r#"
-                    struct User {
-                        first_name: String,
-                        last_name: String,
-                        age: u32,
-                    }
-                    …
-                "#},
-            ),
-            (
-                indoc! {r#"
-                    trait «FooProvider» {
-                        const NAME: &'static str;
-
-                        fn provide_foo(&self, id: usize) -> Foo;
-
-                        fn provide_foo_batched(&self, ids: &[usize]) -> Vec<Foo> {
-                             ids.iter()
-                                .map(|id| self.provide_foo(*id))
-                                .collect()
-                        }
-
-                        fn sync(&self);
-                    }
-                    "#
-                },
-                indoc! {r#"
-                    trait FooProvider {
-                        const NAME: &'static str;
-
-                        fn provide_foo(&self, id: usize) -> Foo;
-
-                        fn provide_foo_batched(&self, ids: &[usize]) -> Vec<Foo> {
-                    …
-                        }
-
-                        fn sync(&self);
-                    }
-                "#},
-            ),
-        ];
-
-        for (input, expected_output) in table {
-            let (input, ranges) = marked_text_ranges(&input, false);
-            let buffer =
-                cx.new(|cx| Buffer::local(input, cx).with_language(Arc::new(rust_lang()), cx));
-            buffer.read_with(cx, |buffer, _cx| {
-                let ranges: Vec<Range<Point>> = ranges
-                    .into_iter()
-                    .map(|range| range.to_point(&buffer))
-                    .collect();
-
-                let excerpts = assemble_excerpts(&buffer.snapshot(), ranges);
-
-                let output = format_excerpts(buffer, &excerpts);
-                assert_eq!(output, expected_output);
-            });
-        }
-    }
-
-    fn format_excerpts(buffer: &Buffer, excerpts: &[RelatedExcerpt]) -> String {
-        let mut output = String::new();
-        let file_line_count = buffer.max_point().row;
-        let mut current_row = 0;
-        for excerpt in excerpts {
-            if excerpt.text.is_empty() {
-                continue;
-            }
-            if current_row < excerpt.point_range.start.row {
-                writeln!(&mut output, "…").unwrap();
-            }
-            current_row = excerpt.point_range.start.row;
-
-            for line in excerpt.text.to_string().lines() {
-                output.push_str(line);
-                output.push('\n');
-                current_row += 1;
-            }
-        }
-        if current_row < file_line_count {
-            writeln!(&mut output, "…").unwrap();
-        }
-        output
-    }
-
-    fn rust_lang() -> Language {
-        Language::new(
-            LanguageConfig {
-                name: "Rust".into(),
-                matcher: LanguageMatcher {
-                    path_suffixes: vec!["rs".to_string()],
-                    ..Default::default()
-                },
-                ..Default::default()
-            },
-            Some(language::tree_sitter_rust::LANGUAGE.into()),
-        )
-        .with_outline_query(include_str!("../../languages/src/rust/outline.scm"))
-        .unwrap()
     }
 }
